Remove unnecessary cast

+++ rrrc/components/SensorPortHandler/SensorPortLibraries/HC_SR04/HC_SR04.c
@@ -142,7 +142,7 @@ SensorLibraryStatus_t HC_SR04_Update(SensorPort_t* sensorPort)
             libdata->isMeasuring = false;
             update_filtered_distance(libdata);
             
-            uint32_t cm = (uint32_t) _get_cm(libdata->filtered_distance_tick);
+            uint32_t cm = _get_cm(libdata->filtered_distance_tick);
             SensorPort_Write_PortState(sensorPort->port_idx, (uint8_t*) &cm, sizeof(cm));
             
             /* mark next sensor as active */
